feat: add file-system specific safeload_yaml
